refactor: split createPolicyFile into focused files
- createPolicyFile.go (89 lines) — Cobra command + file writing
- policyWizardModel.go (133 lines) — bubbletea Init/Update/View
- policyWizardForms.go (432 lines) — step enums, form builders,
  state transitions, form result processing
- policyWizardContext.go (126 lines) — API fetching, styles

Also extracted confirmForm() and inputForm() helpers to reduce
repetition in the form builder switch.
