+++ cmd/kosli/policyWizardModel.go
@@ -1,0 +1,133 @@
+package main
+
+import (
+	"strings"
+
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/huh"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/kosli-dev/cli/internal/policy"
+)
+
+const formWidth = 55
+
+type policyWizardModel struct {
+	step   wizardStep
+	form   *huh.Form
+	policy *policy.Policy
+	wctx   *wizardContext
+	styles wizardStyles
+	width  int
+	height int
+
+	// State for loops and expression building
+	exprTarget     exprTarget
+	exprMode       string
+	exprContext    string
+	exprTagKey     string
+	currentAttRule policy.AttestationRule
+	cancelled      bool
+	outputFile     string
+	requireProv    bool
+	requireTrail   bool
+	validationErr  string
+}
+
+func newPolicyWizardModel(wctx *wizardContext) policyWizardModel {
+	m := policyWizardModel{
+		step:   stepProvConfirm,
+		policy: policy.NewPolicy(),
+		wctx:   wctx,
+		styles: newWizardStyles(),
+		width:  120,
+	}
+	m.form = m.buildForm()
+	return m
+}
+
+func (m policyWizardModel) Init() tea.Cmd {
+	return m.form.Init()
+}
+
+func (m policyWizardModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case tea.WindowSizeMsg:
+		m.width = msg.Width
+		m.height = msg.Height
+	case tea.KeyMsg:
+		if msg.String() == "ctrl+c" {
+			m.cancelled = true
+			return m, tea.Quit
+		}
+	}
+
+	form, cmd := m.form.Update(msg)
+	if f, ok := form.(*huh.Form); ok {
+		m.form = f
+	}
+
+	if m.form.State == huh.StateAborted {
+		m.cancelled = true
+		return m, tea.Quit
+	}
+
+	if m.form.State == huh.StateCompleted {
+		m.processFormResults()
+		if m.validationErr != "" {
+			m.form = m.buildForm()
+			return m, m.form.Init()
+		}
+		m.advanceStep()
+		if m.step == stepDone {
+			return m, tea.Quit
+		}
+		m.form = m.buildForm()
+		return m, m.form.Init()
+	}
+
+	return m, cmd
+}
+
+func (m policyWizardModel) View() string {
+	if m.cancelled || m.step == stepDone {
+		return ""
+	}
+
+	s := m.styles
+	fw := formWidth
+	available := m.width - s.base.GetHorizontalFrameSize()
+	pw := available - fw - 2
+	if pw < 30 {
+		pw = 0
+	}
+
+	header := s.title.Render("Kosli Policy Builder")
+
+	formContent := m.form.View()
+	if m.validationErr != "" {
+		errStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("#FE5F86")).Bold(true)
+		formContent = errStyle.Render("⚠ "+m.validationErr) + "\n\n" + formContent
+	}
+	formView := lipgloss.NewStyle().Width(fw).Render(formContent)
+
+	var body string
+	if pw > 0 {
+		yamlBytes, _ := m.policy.ToYAML()
+		yamlStr := strings.TrimRight(string(yamlBytes), "\n")
+		if yamlStr == "" {
+			yamlStr = "(empty)"
+		}
+		previewContent := s.previewText.Render(yamlStr)
+		previewTitle := s.accent.Bold(true).Render("Live Preview")
+		previewPanel := s.preview.Width(pw).
+			Render(previewTitle + "\n\n" + previewContent)
+
+		body = lipgloss.JoinHorizontal(lipgloss.Top, formView, "  ", previewPanel)
+	} else {
+		body = formView
+	}
+
+	footer := s.footer.Render("ctrl+c to cancel • enter to confirm")
+
+	return s.base.Render(header + "\n\n" + body + "\n" + footer)
+}
